+++ src/plugin/gcollapse.c
@@ -5,7 +5,7 @@
  * Updated: Sat May 20 14:06:47 EDT 2017
  * Purpose: Stata plugin to compute a faster -collapse-
  * Note:    See stata.com/plugins for more on Stata plugins
- * Version: 0.3.2
+ * Version: 0.3.3
  *********************************************************************/
 
 #include "gcollapse.h"
@@ -90,53 +90,81 @@ int sf_collapse (struct StataInfo *st_info)
     // Read in variables from Stata
     // ----------------------------
 
+    int rct = 0;
+
+    /*********************************************************************
+     *                 Method 1: Continuously from Stata                 *
+     *********************************************************************/
+
     /*
-     *
      * The following maps the C group index to Stata so we can read
-     * observations from Stata in order; however, this was not faster,
-     * in testing, than reading observations out of order
-     *
-     * size_t *index_st = calloc(st_info->N, sizeof *index_st);
-     * for (j = 0; j < st_info->J; j++) {
-     *     start  = st_info->info[j];
-     *     end    = st_info->info[j + 1];
-     *     offsets_buffer[j] = start * st_info->kvars_source;
-     *     for (i = start; i < end; i++)
-     *         index_st[st_info->index[i]] = j;
-     * }
-     *
-     * offset_buffer = offset_source = 0;
-     * for (i = 0; i < st_info->N; i++) {
-     *     j     = index_st[i];
-     *     start = st_info->info[j];
-     *     end   = st_info->info[j + 1];
-     *     nj    = end - start;
-     *     offset_buffer = start * st_info->kvars_source;
-     *     offset_source = j * st_info->kvars_source;
-     *     for (k = 0; k < st_info->kvars_source; k++) {
-     *         // Read Stata in order
-     *         if ( (rc = SF_vdata(k + st_info->start_collapse_vars, i + st_info->in1, &z)) ) return(rc);
-     *         if ( SF_is_missing(z) ) {
-     *             if (i == start)   all_firstmiss[offset_source + k] = 1;
-     *             if (i == end - 1) all_lastmiss[offset_source + k]  = 1;
-     *         }
-     *         else {
-     *             // Read into C in order as well, via index_st, so
-     *             // non-missing entries of given variable for each group
-     *             // occupy a contiguous segment in memory.
-     *             all_buffer [offset_buffer + nj * k + all_nonmiss[offset_source + k]++] = z;
-     *         }
-     *     }
-     * }
-     * free (index_st);
-     * if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5.1: Read in source variables");
+     * observations from Stata in order; this is only sometimes faster
+     */
+    size_t *index_st = calloc(st_info->N, sizeof *index_st);
+    for (j = 0; j < st_info->J; j++) {
+        start  = st_info->info[j];
+        end    = st_info->info[j + 1];
+        offsets_buffer[j] = start * st_info->kvars_source;
+        for (i = start; i < end; i++)
+            index_st[st_info->index[i]] = j;
+    }
+
+    offset_buffer = offset_source = 0;
+    for (i = 0; i < st_info->N; i++) {
+        j     = index_st[i];
+        start = st_info->info[j];
+        end   = st_info->info[j + 1];
+        nj    = end - start;
+        offset_buffer = start * st_info->kvars_source;
+        offset_source = j * st_info->kvars_source;
+        for (k = 0; k < st_info->kvars_source; k++) {
+            // Read Stata in order
+            if ( (rc = SF_vdata(k + st_info->start_collapse_vars, i + st_info->in1, &z)) ) {
+                rct = rc;
+                continue;
+            }
+            if ( SF_is_missing(z) ) {
+                if (i == start)   all_firstmiss[offset_source + k] = 1;
+                if (i == end - 1) all_lastmiss[offset_source + k]  = 1;
+            }
+            else {
+                // Read into C in order as well, via index_st, so
+                // non-missing entries of given variable for each group
+                // occupy a contiguous segment in memory.
+                all_buffer [offset_buffer + nj * k + all_nonmiss[offset_source + k]++] = z;
+            }
+        }
+    }
+    free (index_st);
+    if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5.1: Read source variables");
+    if ( rct ) return (rct);
+    /*
      */
 
+    /*********************************************************************
+     *                      Method 2: Out of order                       *
+     *********************************************************************/
+
+    /* For debugging
+     * 
+    for (j = 0; j < st_info->J * st_info->kvars_source; j++)
+        all_nonmiss[j] = 0;
+     *
+     */
+
+    /*
+     * The following reads the variables out of order. It consumes ever
+     * so slighly less memory and was easier to code. I am not 100% when
+     * this is faster than method 1...
+     *
     offset_buffer = offset_source = 0;
     for (j = 0; j < st_info->J; j++) {
         start  = st_info->info[j];
         end    = st_info->info[j + 1];
         nj     = end - start;
+        offset_buffer = start * st_info->kvars_source;
+        offset_source = j * st_info->kvars_source;
+
         // Loop through group in sequence
         for (i = start; i < end; i++) {
             sel = st_info->index[i] + st_info->in1;
@@ -156,17 +184,17 @@ int sf_collapse (struct StataInfo *st_info)
             }
         }
         offsets_buffer[j] = offset_buffer;
-        offset_buffer    += nj * st_info->kvars_source;
-        offset_source    += st_info->kvars_source;
     }
     if ( st_info->benchmark ) sf_running_timer (&timer, "\tPlugin step 5.1: Read in source variables");
+     *
+     */
+
+    // Collapse variables by group
+    // ---------------------------
 
     for (j = 0; j < st_info->J; j++)
         for (k = 0; k < st_info->kvars_source; k++)
             nmfreq[k] += all_nonmiss[j * st_info->kvars_source + k];
-
-    // Collapse variables by group
-    // ---------------------------
 
     // We encoded stat string; see mf_code_fun in gtools_math.c
     for (j = 0; j < st_info->J; j++) {