+++ src/plugin/gtools.h
@@ -57,5 +57,6 @@ struct StataInfo {
 int  sf_parse_info  (struct StataInfo *st_info, int level);
 int  sf_hash_byvars (struct StataInfo *st_info);
 void sf_free        (struct StataInfo *st_info);
+int  sf_numsetup    ();
 
 #endif